requirements.txt
# Target runtime: Python 3.11 (compatible with 3.10+)
torch>=2.1.0,<3.0.0
torchvision>=0.16.0,<1.0.0
lightning>=2.2.0,<3.0.0
anomalib==2.3.3
opencv-python-headless>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
streamlit>=1.28.0
scikit-learn>=1.3.0
Pillow>=10.0.0
PyYAML>=6.0.0
pytest>=7.4.0
